update to model
