+++ src/gui/HomeScreen.java
@@ -11,8 +11,8 @@ import javax.swing.JOptionPane;
 import java.awt.FlowLayout;
 import javax.swing.SwingConstants;
 
-import classes.User;
 import client.Client;
+import model.User;
 
 import java.awt.Font;
 import java.sql.Date;
